Change the default tag-to-hostname template.
**This is a potentially breaking change.**  In 0.6 we used the effective template
`{{.Name}}-{{.Tag}` for turning Route tags into K8s services.  Unfortunately, when
used in conjunction with the BYO-revision-name feature there is a very natural
collision scenarios:

```
spec:
  template:
    metadata:
      name: prefix-v1234
    spec: ...
  traffic:
  # v1234 is a very natural choice here
  - name: v1234
    percent: 12
  ...
```

Unfortunately, we create a Service for the revision named `prefix-v1234` *and* a
"placeholder" service for the Route tag with `{{.Name=prefix}}-{{.Tag=v1234}}` so
only one can win and we end up with a `NotOwned` failure of some sort because the
two controllers fight.

**To preserve the 0.6 behavior** you can simply update `config-network` to specify:

```
data:
  tagTemplate: "{{.Name}}-{{.Tag}}"
```

This changes the default to `{{.Tag}}-{{.Name}}` to avoid the `{{.Name}}-` prefix,
which is required by BYO-revision-name.

Fixes: https://github.com/knative/serving/issues/4206