Fix the pod counting target.
We need to count the number of pods in the actual revision's metric service.
This way were counting pods in the revision target, which is always 1, since
the activator is hooked in.
Obviously this caused some problems like:

'Failed to scrape metrics,error:fail to get a successful scrape for 1 tries: Get http://s4-v77zd-metrics.default:9090/metrics: dial tcp 10.27.247.15:9090: connect: connection refused'