Alternative solution for scrape target service update.
In this case we update the metric scraper when the metric changes.
I am not a fan of this solution, but keeping scraper immutable has other benefits,
including not requiring to pass in the lister to the core, as Jaques suggestes in #4303.

But, all in all, I am fine merging this (and then backing off the changes I did for scraper URL updates).

/cc @mattmoor @jchesterpivotal