Factor out ScaleResource related methods. (#4334)
These "helper" methods have been hardwired into the scaler. The scaler however is only used for the KPA case, whereas these helper functions are needed generally to resolve any PodAutoscaler's scaleTargetRef into the actual object. That is needed for example to create the services in the ServerlessService and to create the MetricService from the PodAutoscaler by setting the correct selector.

This is part of the refactoring needed to provide a base reconciler for both HPA and KPA.