Create a new autoscaling base reconciler. (#4351)
* Create a new autoscaling base reconciler.

HPA and KPA share a lot of setup logic and code so this factors out a general autoscaling Base reconciler which contains the necessary bits that are needed for both.

Note that this already take the future of the HPA into account in that it also contains the functions to fully reconcile the metrics service and to generate the Metrics entity to trigger metric collection.

* Capitalize abbreviated names.