Have v1beta1 test use typed client.
Originally this used the dynamic client because we hadn't generated clients, but now that we have them we can use them for this half of the v1beta1 test.  The other path still needs the dynamic client in order to test sending a payload that cannot be properly formed through the typed client.