Revise PodAutoscaler's life cycle
Currently PodAutoscaler's Ready condition is the same as Active
condition. It is not only mismatched with Revision's Ready condition,
where the Active condition is only an informative condition that does
not define the Ready condition, but also not capturing the
PodAutoscaler's status very well. For example, when everything is good
but there's no traffic, the PodAutoscaler should be Ready, but Inactive.

Changes are:

1. Make Active an informational condition for PodAutoscaler.
2. Define ContainersHealthy and PodsHealthy as living conditions for
PodAutoscaler's Ready.
3. PodAutoscaler checks pods and determines PodsHealthy and
ContainersHealthy conditions.
4. Revision reads PodAutoscaler's PodsHealthy and ContainersHealthy
conditions and updates its ContainerHealthy and ResourceAvailable
conditions. Note that Revision can have extra checks for those
conditions.
5. PodAutoscaler scales down (to 0 if possible) when it fails to become
Ready.