Fix the percentage to actually be percentage (#4413)
* Fix the percentage to actually be percentage

The value of the container-concurrency-target-percentage is a fraction, which is
contradicting the name of the key.
So fix it to be %, but permit backwards compatibility by adjusting the values in the service code.
Values close to 1% seem meaningless anyway.

Also add validation to the value of said variable.

In general all values should be sanity checked.

* also rename the field internally, since internally we're keeping fractions