Allow custom-metrics server to generate certs as a non-root user.
If the custom-metrics server is not given any certs via the command-line (which is our default behavior for now) it will generate and serve them on the fly. The default directory for this is `apiserver.local.config`, which a non-root user cannot create.

This switches the directory for these throw-away certificates to `/tmp` to make the autoscaler runnable as a non-root user again.