Filter all endpoints informer by the label we're looking for.
The `EnqueueLabelOfNamespaceScopedResource` function itself filters as well, but a bit too late. We'll still see a "reconcile succeeded" log for non-managed endpoints, which is noisy and misleading.

This adds top-level filtering to the event-handler to make sure that endpoints that are not managed by Knative don't even reach the controller implementation.