Make Target() return float64 (#4445)
* Make Target() return float64

I had wanted to this change earlier, when the only usage was in KPA, where we cast
it to float.
But now it's in use in HPA as well, so the change is not as clear-cut as before.
I still think it's useful, and FWIW, we might want to permit non whole CC  :-)

/assign @markusthoemmes

* ceil in the HPA