Fix the bug where we might have ended up with 0 target concurrency.
The code as it was, would multiply target currency by target utilization.
Since TU is effectively 1, this was never a problem.
But now that we're trying to lower it, it could have been that we'd have numbers <1, which after int() casting would
effectively become 0.
This fixes the problem in both places (the CFGMAP setting) and the computation (min of 1, will be ensured)>

/assign @mattmoor @markusthoemmes