Redo 'TestDestroyPodTimely' to narrow its area under test.
This test used to delete the whole service it created to in turn check that the pod that eventually gets created by the service through a cascade of objects gets deleted in a timely manner.

It's now narrowed down to check precisely **1** pod of the created service, delete it and watch the state of exactly that pod until it's successfully shut down.