Make the 'AutoscaleUpDownUp' test more stable. (#4488)
* Make the 'AutoscaleUpDownUp' test more stable.

This test used to use a specific helper to determine the autoscaler scales up correctly. While a lot of work has gone into other helpers to deflake things, this code has been sitting there for quite a while, untouched.

This uses the more common helpers to assert scale in that test as well. They are more stable and account better for the reality of metric scraping in the system.

* Remove unneeded function.