Relax 'TestMultiScalerTickUpdate' to make it more stable.
This test had a very narrow timeout setting (poll once per millisecond for 10 milliseconds), which is usually fine but gets hairy on non-dedicated test machines.

It now uses the same timeout parameters all the other tests in this package use as well, so they can be adjusted more globally in the future. The diluted timeout settings do not compromise the value of the test.