Switch the const to be 0.7 for target concurrency utilization (#4420)
* Fix the percentage to actually be percentage

The value of the container-concurrency-target-percentage is a fraction, which is
contradicting the name of the key.
So fix it to be %, but permit backwards compatibility by adjusting the values in the service code.
Values close to 1% seem meaningless anyway.

Also add validation to the value of said variable.

In general all values should be sanity checked.

* also rename the field internally, since internally we're keeping fractions

* Revert default to be 0.7 for the concurrency target

Fixes #2016
Stop running at 100% for the concurrency target, but rather at 70%, by default.

/hold
/assign @mattmoor @markusthoemmes

* remove markers

* more comment removal

* Adjust to the new target utilization

* fix stuff

* use config map for the constant