Replace timeout propagation with context propagation in breaker.
The use-case of potentially aborting a long-running function like the breaker's `Maybe` seems like a perfect fit for the use of Contexts. This has the added benefit that the tests are no longer timing based and that an abortion of the outer HTTP request actually aborts trying to obtain a semaphore.