Update Runtime Contract with links to Runtime tests (#4428)
Add links to the associated tests for tested statements in the runtime
contract. To avoid frequently changing this the link points to the test
file itself instead of the exact line or function within the test.

This should make it easier to visually see tested statements as well as
quickly look up how a particular statement is being tested.