Add simple probe cache for the activator. (#4618)
* Add simple probe cache for the activator.

This is a simple probe cache for the activator.
With the target capacity work being introduced we are putting
a lot more traffic via Activator. When we're running the system like this
there is no reason for us to actually probe revision on every request.

The work @greghaynes is doing to make it even more generic (e.g. to listen
to the deployment size) and take it to the next level.

/assign @mattmoor @markuesthoemmes

* review issues

* struct{}

* more stuff

* test consts