Stop relying on timers in TimeToFirstByteTimeoutHandler's tests.
This test relied on timers to hit a timeout reliably. The values of these timeouts are a compromise between stability of the test and runtime of the test.

This removes the needs for these timers in almost all of the tests and instead relies on locking and waits for however long it takes for signals to propagate through the system.