Redo breaker tests to ensure consistency and to not rely on polling. (#4646)
* Redo breaker tests to ensure consistency and to not rely on polling.

The breaker tests used to rely on a mechnism to poll the internal state of the breaker to eventually become consistent. In a concurrent setting, this is prone to errors and subtle bugs.

This is a complete overhaul of these tests. It relies on the fact that failed requests terminate immediately and that this failure can be used as a mechnism to detect whether the internal state is consistent or not. Essentially, if a request fails it is safe to say that the breaker is at capacity.

* Fix typo.

* Remove unneeded tests, test renaming, added specific queueing test.