Add an explicit test for the number of probes in the activator. (#4642)
We implicitly tested the number of probes in tests that run requests concurrently. As such, we cannot safely guarantee the number of probes. In the worst case, the probes could've been run on all of the requests.

Instead, we can test the probes explicitly in a test that runs one request at a time. That way we know for sure how many probes to expect and can test accordingly.