Some minor optimizations to how we format host:port strings. (#4652)
In most cases we can avoid sprintf, which only for the means of formatting integer runs 4x slower than atoi,
pile in additional memory for parsing the request string, etc.

So do just that.

Also some ports were always used as strings, so define the constant as string, rather than as number.