Exclude the custom-metrics API port from istio's interception.
When a sidecar is injected, istio intercepts all application ports by default. In the case of the custom-metrics API, the Kubernetes API server will try to talk to the endpoint and fail because of SSL certificate issues.

Excluding the custom-metrics port from istio's interception helps alleviate the issue and is probably a good idea anyway.