Fix dep hanging for a very long time without mercurial. (#4712)
If mercurial is not installed, `dep ensure` can hang for a very long time because it tries first to clone mercurial repos via `hg` (which fails), but in the case for bitbucket.org based repositories it'll also try to clone it via git, which eventually fails as well but has a very long timeout.

This gives dep a hint that this is actually a mercurial repository, so it won't try git at all and fail straight away.