Remove use of python image for CmdArgs test (#4721)
The previous python image that we used came from Dockerhub while all
other test images are uploaded to the registry of choice by the test
runner. This removes the use of python so that all tests use images that
we upload. This gives us more control and flexibility.

This change also moves CmdArgs to runtime instead of api. While we do
not have anything about Cmd or Args in our runtime contract moving it
here seems more appropriate as we are validating that the runtime
environment recieved and expressed the correct behavior. We are not
evaluating that the control plane expresses the correct status.