Activator PodIPWatcher
This creates a new PodIPWatcher which tracks the health of a revision's
PodIPs asynchronously and allows for getting a list of healthy endpoints
in our request path with a single read lock. This will be expanded to
include tracking of a Revision's private service ClusterIP as well and
will return the appropriate dest depending on its health.

Related-to: #3885