Make breaker tests deterministic again. (#4784)
It's very important in these tests to assert failing requests to make sure the breaker's state is consistent. The change caused the "to be canceled" request to race the "supposed to be successful" request before it to the semaphore. If the "to be canceled" request won, it'd get capacity immediately and the cancel is defunct. The "supposed to be successful" request before it gets put into the queue and the `expectFailure` call hangs forever.