Stop relying on naming conventions in the activator's throttler logic. (#4796)
The throttler code needs to get the revision's name based on endpoints floating in to update it's internal capacity counts which are per revision. That lookup is based on a naming convention currently. The throttler assumes that the endpoint's name is the revision's name appended with a random string delimited by a dash. A typical generateName schema.

generateName however is [smart enough](https://github.com/kubernetes/kubernetes/blob/7f23a743e8c23ac6489340bbb34fa6f1d392db9d/staging/src/k8s.io/apiserver/pkg/storage/names/generate.go#L42-L54) to know that some resources have a limit of 63 characters. It makes sure that this limit is not surpassed thus breaking the assumption taken above for very long names (as is the case for the service-to-service tests for example).

This drops that assumption and instead relies on a label that's already propagated through to the endpoints.