Deeply replace the 'MetricKey' with 'NamespacedName'. (#4894)
The metric keys are currently a concatenated version of namespace/name. However, working with strings all throughout the codebase is kinda brittle as in the end, anything could be in them.

A much more typesafe method is to use 'NamespacedName' from k8s apimachinery. As it is only a struct with two strings, it is comparable via == and thus can even be used as keys for maps.