Bump number of retries to 15
We're still seeing timeouts in dialing.
Those are traced back to IP table changes in k8s, which are, e.g. on GKE
are staggered every 10s.
This is controlled by .
The per-pod probing and proxying should alleviate this problem,
but this will in the meantime remove the errors in the logs.