Introduce the metric controller.
The metric controller reacts on the lifecycle of metric objects and currently makes sure that metric collecting is started when a metric is created and stopped if one is deleted. As updating is not really a distinct use-case we want to handle differently, this also introduces a 'CreateOrUpdate' function to the collector, which appropriately acts if the collection was already running or not.

This is a stepping stone towards removing the in-memory handling of metric entities. The controller is not yet hooked up though.