Optimize throttler for cases where the breaker is already present. (#4912)
* Optimize throttler for cases where the breaker is already present.

`getOrCreateBreaker` checks whether a breaker is already present. If so, it returns it. If not, it creates a new one. To do so though, it needs to acquire a write-lock. Always acquiring that write-lock results in a lot of contention and other requests, even if they go to an entirely different revision, are excluded from that path.

I investigated the usage of a concurrent map for this use-case and while that might still be a valuable thing to do, it comes with a few caveats. Most notably, all concurrent maps I could find have no method to lazily compute a value if it's not yet set. Since we atomically need to get-or-set (we only want to create one instance of the breaker) we'd need to generate the new breaker on every call. Not nice.

Instead, this relies on an optimistic approach. It acquires a read-lock first and tries to get a breaker. If no breaker is present, we get a write-lock and create a new breaker after checking, if we haven't been raced. That reduces the contention in the happy case (which is the vast majority of proxying requests) to just read-locks.

* Give the pattern a name.