Update the PodAutoscaler status when desiredScale is unknown
Currently, the PA status is left alone when the `desiredScale` is
unknown and the current scale is less than `minScale` (default = 1).

This is inaccurate in light of #4183, which introduces the possibility of
the `minScale` changing during the lifetime of the PA. Tangentally, this
inaccuracy is the source of some flakes in the `MinScale` test in #4484.

This change marks the PA as activating whenever the `desiredScale` is
non-zero and the current scale is less than `minScale`.