Reduce the possible minimal grace period.
With the latest changes we've done to the scale to zero behavior, especially in the face of the
Activator being in the path for TBC work, it is possible to have quite short Grace windows.
This change permits it to be as small as minimal window.

This should help everyone who wants faster shutdowns.

For #4453

The integration test to test all this is in progress...