Move definition of concurrency types
The motivation for this change is to make it possible for
pkg/apis/serving/... to depend on pkg/apis/autoscaling.

I'd like to migrate our reconcilers to use this pattern:

parent.PropagateChildStatus(child)

But having these constants defined in serving means that we can't do
that for the Revision -> Autoscaler relationship. Regardless, this seems
like a better place for these constants to live, since they control the
autoscaler behavior.