Fix flaky unit test
This waits for informer caches to sync after creating the revision and
before creating the route. By doing this, we can ensure that the
informer will be populated with the revision when we reconcile the
route. If the revision doesn't exist when we reconcile the route, we
currently don't bother tracking the revision via the tracker, so we
never re-reconcile the route once the informer syncs. This feels like
dropping updates, but I'm not sure if this race condition ever happens
in practice.