Have the labeling reconciler label Revisions. (#5048)
* Bring back the application of `serving.knative.dev/route` to Revisions.

We used to do this, and we stopped for a few reasons.  Based on my recollection
it was some combination of complexity in the Route controller (before this was
a separate Reconciler) and reconciliation churn (before we had things under
better control).

This simply clones the Configuration logic and replaces strings.  I wanted to do
this in a separate commit (with passing tests) before refactoring to deduplicate
the otherwise identical code for Configuration/Revision.

* Refactor the labeling Reconciler to share its core logic.

This adds abstractions to deduplicate the logic from the previous commit by
making it go through an "accessor" abstraction, that we implement for revisions
and configurations.

* rename test function