Ensure that the Route label isn't propagated to the Deployment.
We recently started to apply the serving.knative.dev/route label to
Revisions in addition to Configurations.  I noticed that when I was
deploying things there was higher pod churn then expected, and this
was due to the fact that when the Revision became routable and the
label was applied a rollout of the Deployment with the new label
was triggered.

This simply filters the label from those we propagate down to the
Deployment.