Refactor deployment status propagation (#5077)
This change brings the Deployment -> Revision status propagation more in
line with how other reconcilers work. Since this is at the edge between
knative and kubernetes resource conventions, we need to transform the
deployment conditions to conform to our conventions, e.g. by inverting
appsv1.DeploymentReplicaFailure to be a happy condition and exposing a
top-level happy state ("Ready") for the deployment.

There are some changes to behavior:

- An event is no longer emitted when the deployment times out.
- We surface the underlying DeploymentProgressing Reason/Message instead
of hard-coding our own.
- We surface the DeploymentReplicaFailure message as well.