Rely on a given point in time for 'inStatusFor' checks.
We all know relying on the global clock is kinda bad and will at some point drop on your feet. This happened here in that the 'ActiveFor' tests are flaky because it checks against a fixed value, which can never be guaranteed and will be skewed if the CPU scheduler goes out for lunch in between executing two statements under test.

Instead of making the test accept diluted values, I opted to actually inject the point in time to check against to make it reliable in tests.