Add `Reachability` field to PodAutoscaler spec (#5147)
* Add `Reachable` field to PodAutoscaler spec

This field can be used by the KPA and HPA to make scaling decisions
based on whether the `ScaleTargetRef` is "reachable" (ie. has an active
route).

The rationale is to abstract away the implementation of reachability in
Serving from the decisions autoscaler needs to make based on whether or
not the `ScaleTargetRef` is reachable (eg. #4183).

(extracted from #4884)

* Autogen

* Add comments to explain possible values

* Expand on pa.IsReachable() comments

* Convert *bool to enum

* Unclobber suggested edits