This create a simple deployment latency prober. (#5267)
* This create a simple deployment latency prober.

The basic idea is: For X minutes create a service every Y seconds and record how long
it takes to become ready (don't delete any until the end).  I have experimented with
this at a variety of scales, in order to explore how many services we can support and
seen some interesting results so far.

Here is an example with X=55,Y=5 where we successfully create ~650 services that become
ready, but which exhibit some clearly pathological behaviors we will need to track down:
https://mako.dev/run?run_key=4727463267532800&~dl=1&~cl=1&~rl=1&~rvl=1&~il=1&~sksl=1&~pal=1

I am checking this in with X=25,Y=5 so that we can get multiple datapoints per cycle, and
this is what my cluster is currently configured to run, so I should have some interesting
data in a few more hours, which will allow us to start thinking about what goes into
`sla.go`.

* Address review feedback

* Add analyzers with thresholds derived from where we are.

* Convert duration and frequency to flags.