Run the deployment prober for longer.
This benchmark originally ran twice an hour for 25 minutes, but one run was being truncated by the hourly serving install, which aborts in-progress jobs, which lead to reporting the partial data for the run, which consistently tripped our analyzers.  Now that this runs once an hour, run it for longer.