Fix wrong cat and kubectl usage in test script
This command does not input anything to `cat` command.

```
cat | kubectl apply -f - <<EOF
```

Hence, `kubectl apply` works fine, but we need to push `Enter` key to
finish cat command and proceed the script.

To fix it, we need either

```
kubectl apply -f - <<EOF
```

or

```
cat <<EOF | kubectl apply -f - <<EOF
```

This patch changes the script to the latter, which is same style with
https://kubernetes.io/docs/reference/kubectl/cheatsheet/.