In the autoscaler store decider ptrs (#5366)
* In the autoscaler store decider ptrs

In the autoscaler currently we keep the actual decider objects and we copy them quite often
and it is not the smallest object we have in stock.

On the other side, the autoscaler does not mutate the object, i.e. it's immutable outisde of
KPA.
So let's deep copy once and store a copy in uni/multi-sclares.

In addition we were creating a copy of it on every Autoscaler.Scale run, which is a problem in itself.

* fix isseu