Still sendUpdate from checkDests after ClusterIP is up. (#5417)
Previously we elided the sendUpdate when the ClusterIP had been successfully probed, but this had the side effect that backend capacity of the revisions wasn't updated until either:
1. The revision scales to zero, or
1. The activator restarts.

This is a problem for both increasing and decreasing the capacity of the Revision when the activator is on the request path.

The change is to simply elide the probe, but still sendUpdate with the same payload we'd send if we had just probed successfully.

With this change (and lots of omitted logging) I was able to confirm that cases where we previously ended up with inadequate capacity now have the appropriate target capacity for throttling in the case I was using to reproduce the problem.

This also removes a test that this change breaks because correcting it reveals a race in the test logic (likely around the sleep), and I'd like to get this change baking in the dataplane-probe.