Add sleeps to the tests that panic in the logger.
Currently we have no way to knowing when the informers stop. And if that's even possible.
So these are bandaids for the tests that are now failing a lot on the prow.
I've verified all of them with -race -count=5 setup on my machine,
but let's see how they work on Prod.
I don't really want to put the sleep in the cancel() since a lot of
tests don't need it.

Longer term, we need to find a way to find out whether the informer stopped actually doing stuff
or do something with logger so that it knows test terminated and does not pass on to t.Logf.

/assign mattmoor