Fix 500s during cold start.
We were seeing the impossible!

```
  error:  "made it through breaker but we have no clusterIP or podIPs. This should never happendefault/load-test-200-8pckm"
```

After tracing through the logic, we properly set capacity to zero when things scale to zero, but when cold starting we would update the revision's "capacity" based on the number of ready addresses in the endpoint before we had successfully probed it.  This led to the condition above, where the breaker lets traffic through, but has nowhere to send it.

The change is relatively simple, we wrap the call to `updateCapacity` in a check that we have somewhere to send things.
 * If we do, we pass through the actual backend count,
 * If we don't, we pass through zero.

In my dev setup this seems to eliminate the 500s in its first few runs.