Always pass Healthy dests to the throttler
TL;DR: This is needed for future "ideal" load balancing in throttler wherein I
will need the actual pods to send traffic to -- so always send the ready
pod IPs. But this doubles as `ReadyPodCount`, so remove that.

As discussed with Greg, but for posterity:

Right now we know everything that is in dests is healthy. So when we switch to clusterIP
we can forward all the pods there for endpoints count reasons.

In future when we will have both healthy and unhealthy pods passed into the
we can combine and probe that combined list and send all the healthy ones to the throttler as we do now.

In any case backendCount==|dests|.

In theory we might probe unready addresses even when we're using cluster IP since we're delegating loadbalancing decision to the throttler,
but this might break the capacity computations, since there might be more pods to use than cluster IP can reach, so this might not be
a good initial approach, but a future enhancement.

/assign @greghaynes @markusthoemmes mattmoor