Add the public interface to the new scale down rate functionality in Autoscaler (#5481)
* Add the public interface to the new scale down rate functionality in autoscaler.

As described in #4993 currently we have no scale down functionality
that matches scaleup rate, but it helps to smooth out sudden sharp short (SSS) drops in traffic
and permits avoiding jitter due to new pod scraping that do not have enough concurrency yet.

This is the public interface to the functionality and the core implementation to follow.
/assign mattmoor @yanweiguo @markusthoemmes

* add better comments

* fix typo