Change error to warning level for revision throttler's update failed log
This patch makes a tiny change, which changes the throttler's update error
log to warning from error level.

Whenever we delete ksvc, `handleUpdate()` outputs error as `Failed to
get revision throttler for revision`. This is no problem as we deleted
ksvc and the revision. Also, `handleUpdate()` is called again, even if
it failed so we should output the message with warning level.