Add a dummy liveness probe to the queue-proxy
This is an alternative fix to what @tcnghia found in https://github.com/knative/serving/pull/5540.

When run with the Istio mesh, Envoy is configured to proxy traffic into the pod and blocks traffic
on any unrecognized ports.  Istio has logic for special casing ports in probes, but apparently not
lifecycle hooks, so when we switched to exec probes the way we shutdown user containers stopped
working with the mesh.

To workaround this, we expose a simple dummy liveness (so that it doesn't affect coldstart) probe
that always returns 200 and serves no purpose other than to trick Istio and workaround the bug.