Split the throttler test in two and make sure we don't depend on order. (#5536)
As we've moved to sets for passing around IP addresses the order of iteration is no longer deterministic
which causes some tests to fail.

Since all we case is actually set comparisons, so I switch to that in the success tests and split out
the failure tests into their own test, where we try for slightly different behaviour so order of events
is important, rather than the IPs where events went.

Also do renames of strings to named constant