Improve the throttler code with injected informers. (#5593)
* Improve the throttler code with injected informers.

This moves throttler along the lines of injectable informers and clients,
which makes the setup of the activator one step closer to being simpler.

There is still work here to deal with a few other components,
but now I can see the line and then we can start working on the actual
controller work.

/assign mattmoor

* better strings

* unwrap