Downward http scheme only when httproto is enabled on autoTLS. (#5201)
* Downward http scheme only when httproto is enabled on autoTLS.

When autoTLS is enabled, the url gets http scheme whenver certificate is not ready.
It is very kind if `httpProtocol: Enabled` is set, but it does not make sense if
`httpProtocol: Disabled` (or `Redirected`) as http is not available.

This patch downwards http scheme only when httproto is enabled on autoTLS.

* Add unit test for Route status check when Certificate is not Ready

* Update to v1 from v1beta1