Fix flake in 'TestKnHTTPSuccessWithThresholdAndFailure'.
This test used to use a "not-equal" condition to check if the request's count matches up with the set thresholds. The prober however returns a success if the success count reaches "greater-than-or-equal" successful requests. The test should just check against a "smaller-than" condition to build the inverse of the prober itself.