Deflake the TestNewConfigurationCallsSyncHandler test (#5681)
This is a new attempt at deflaking the picky TestNewConfigurationCallsSyncHandler test.
Since basically the whole premise of the test is to verify that we create revision when
configuration is reconciled, so let's just do that, without additional movements.
Also fix the strigs across a few tests.
Also move the defer close to the creation of the CB>

/assign @dgerd mattmoor