Update route visibility test.
The original test's intention is to check that the route maintains the
least visibility for the subroutes. It does this by changing both the
route and subroute's visibility to "cluster-local". It then asserts the
route is indeed cluster local then proceeds to update the route object
to be public. While it waits for the state to remain "cluster local".
The test is essentially correct except.. When we ran this against
knative v0.8.1 which doesn't actually implement this, it still passed.

The error stems from the fact that the wait checks for the visibility to
be clusterlocal which it is before the reconciliation finished. Whoops!

So if the test was slow or your cluster is fast, the test will fail
appropriately.

Signed-off-by: Shash Reddy <shashwathireddy@gmail.com>