Only probe gateway endpoints in the same namespace as the gateway itself.
If Istio (or derivatives of it) has multiple control-planes installed in a cluster, there will be multiple gateways, for example 2 istio-ingressgateways. The labels on the service/endpoints for both of these will be identical though. Not qualifying the namespace to look for the service/endpoints pair causes us to pick up ingressgateway pods for both control planes in this example.

This instead qualifies the List/Get operations with the namespace of the respective gateway to make sure that the right resources are returned.