Fix race in activator revisionWatcher
revisionBackendsManager uses revisionWatcher.destsCh both for sending endpoints
and as a signal that a revisionWatcher is done.

Since destsCh is getting closed from the receiving side in revisionWatcher.run(),
this may cause panic when sending on it in revisionBackendsManager.endpointsUpdated().

Fix this by introducing new channel for notifying revisionBackendsManager that a
revisionWatcher is done.

Fixes #5864