Decouple 'TestActivationHandler' from timing to stabilize it.
This test doesn't really **need** to depend on correct timing. As the throttler is mocked away anyway, there is no point in creating a context with a tight timeout only to then sleep in the mocked throttler to trigger said timeout in a test. We might as well just return the resulting timeout error directly.