Prevent duplicate wildcard certificates (#5590)
Previously the nscert reconciler looked for wildcard certs labeled
with the default domain. In the case of changing the domain in the
domain config it would fail to find the existing one and instead
create a new one.

Now the reconciler will check all certs with the label regardless
of the label's value and update the existing one accordingly.

Fixes: #5538