Fix Verbose Logging
In #5398 we missed that flags cannot be set until after they are parsed.
As golang/testing is now doing the parsing just before calling our test function,
to change values they must now be called during the test (or before if we
were to use TestMain -- a future improvement planned). So split the flag setup
into definition and value setting, then call it in the tests that use ServingFlags.

The actual flag reading/parsing has been put into pkg:
https://github.com/knative/pkg/pull/870