Label Configuration/Service as "podspecable".
This also synthesizes a cluster role with each of the podspecable and addresable fragments, so that bindings on podspecable things and addressable resolution may aggregate on these same labels.  With this, we can (eventually) drop some similar logic in knative/eventing, which currently hardcodes serving stuff.

Related: https://github.com/knative/pkg/pull/917