Remove the second level map from metric collection. (#6203)
* Remove the second level map from metric collection.

Now that we're using 1s buckets to collect metrics and now that we're sampling pods for metric collection
we can get rid of the second level map, since we no longer need second degree averaging.

This simplifies the code quite a bit.
Note that while we're using maps, we need to pass pointers along (we were using maps
which are pointers in effect), but this will be removed when we switch to the
array based aggregator.

/assign @yanweiguo @markusthoemmes

* simplify more

* metrics