Parse sub 1% values as fractions.
I have seen users put `0.2` as the value for this annotation, since many users thing in form of fractions.
So instead let's view the values less than 1 as fractions. We already do this for some of the Autoscaling annotations,
where it has been so for historical reasons.
Update the tests, add coverage for the convertor.

/assign mattmoor