Avoid update conflicts in 'TestUpdateConfigurationMetadata'.
This test waited for 'LatestCreatedRevision' to be set before going ahead and updating labels and annotations on the Configuration. This leads to a race with the reconciler as it will eventually also set 'LatestReadyRevision' causing the test to fail with the update.

We now wait for the Configuration to be entirely Ready, which means that it has fully reconciled the created Revision and is now in a steady state.