Make sure to zero out holes if we miss a record.
Currently we only zero out the current element or all of them if we were not collecting data at all for more than a window size.
This is not strictly always correct. Due to the time drift or some other problem we might miss a second and have a whole.
So fix that.
And fix it generally so any sized hole will be zeroed out.
Also update some tests and clean up test file a bit.

/assign mattmoor